+++ src/python/approxeng/input/gui/console.py
@@ -1,19 +1,30 @@
 import curses
+import logging
 import pprint
 from time import sleep
 
 import approxeng.input.sys as sys
-from approxeng.input.controllers import print_devices
+from approxeng.input.controllers import print_devices, find_all_controllers
 from approxeng.input.selectbinder import ControllerResource
 
 
 def sys_scan():
+    logging.basicConfig(level=logging.DEBUG)
     pp = pprint.PrettyPrinter(indent=2, width=100)
     pp.pprint(sys.scan_system())
 
 
 def list_devices():
+    logging.basicConfig(level=logging.DEBUG)
     print_devices()
+
+
+def list_discoveries():
+    logging.basicConfig(level=logging.DEBUG)
+    for discovery in find_all_controllers():
+        logging.info('found a thing')
+        print(discovery)
+        logging.info('done finding a thing')
 
 
 def show_controls():